+++ Logic/Kalender/DsaTool/DSADateCalendarDwarfs.cs
@@ -43,8 +43,8 @@ public class DSADateCalendarDwarfs : DSADateCalendar {
     
     public override String getHeadingText() {
         int jday = getJDay() - 1;
-        int monthdiv = MathUtil.divisio(jday, DAYS_PER_MONTH);
-        int monthmod = MathUtil.modulo(jday, DAYS_PER_MONTH);
+        int monthdiv = (int)MathUtil.divisio(jday, DAYS_PER_MONTH);
+        int monthmod = (int)MathUtil.modulo(jday, DAYS_PER_MONTH);
 
         string s = String.Format("{0:2}. {1:13} ({2}.) {3}", monthmod + 1, monthNames[monthdiv], monthdiv + 1, getYearString());
         return s;
